Add transitions between point scoring
After each time a point is scored, a wipe animation will be played while the objects' positions are reset.

Clicking the Main Menu button on the pause menu will now play a wipe animation before switching scenes.
The transition formerly did not function correctly due to the game being paused, meaning the animation could not be played.
To solve this, before unpausing the game to play the wipe animation, the movement speed of the paddles and the ball are set to 0.