Add settings menu
All of the guis within the settings menu have been laid out. The settings menu can be loaded from the main menu and the "cancel" and "apply" buttons return back to the main menu, all with trans-scene transitions.

The "apply" button needs to be given functionality.
The individual settings need to update upon loading of the settings screen based on the current settings.