Give settings functionality
The settings menu has now been linked to the PlayerPrefs. The setting GUIs are updated to the current PlayerPrefs upon the opening of the menu and the PlayerPrefs update according to the state of the GUIs upon clicking Apply.
A Revert button has been added as well to revert to the original settings.